requirements.txt
streamlit
google-genai
python-dotenv
pytest
pytest-mock
